Extracted advanced node authoring to its own component. #2691
