+++ src/main/webapp/wise5/components/componentController.ts
@@ -343,10 +343,10 @@ class ComponentController {
 
     this.$scope.$watch(() => {
         return this.$scope.$parent.nodeAuthoringController
-            .showAdvancedAdvancedAuthoring[this.componentId];
+            .showAdvancedComponentAuthoring[this.componentId];
       }, () => {
         this.showAdvancedAuthoring = this.$scope.$parent.nodeAuthoringController
-            .showAdvancedAdvancedAuthoring[this.componentId];
+            .showAdvancedComponentAuthoring[this.componentId];
         this.NotificationService.hideJSONValidMessage();
       }, true);
   }
